Time: 0 ms (100.00%), Space: 5.9 MB (28.62%) - LeetHub

+++ 29-divide-two-integers/29-divide-two-integers.cpp
@@ -4,7 +4,7 @@ public:
           if (divisor == -1 && dividend == -2147483648)
              return 2147483647;
 
-        long long result= dividend/divisor;
+        int result= dividend/divisor;
         return result;
     }
 };
